Clases corregido

+++ Clases/src/com/krakedev/Producto.java
@@ -3,6 +3,6 @@ package com.krakedev;
 public class Producto {
 	public String nombre;
 	public String descripcion;
-	public double precio;
+	public float precio;
 	public int stockActual;
 }
